everything that used to work still works more or less

+++ counterfactuals_shapley/compare.py
@@ -2,12 +2,14 @@ import glob
 import lzma
 import os
 import pickle
+import random
 from functools import partial
 
 import numpy as np
 import ray
 import supersuit as ss
 import torch
+import torch.nn.functional as F
 from captum.attr import IntegratedGradients
 from ray.rllib.algorithms.ppo import PPO
 from ray.rllib.env.wrappers.pettingzoo_env import ParallelPettingZooEnv
@@ -59,7 +61,6 @@ def extract_observations(history, agent_name, n, m):
                 obs_n_plus_m.append(data.get("observation")[1:3])
     else:
         return
-
     return obs_n, obs_n_plus_m
 
 
@@ -71,8 +72,16 @@ def extract_pairs_from_histories(histories, agent_name, m, n=None):
         for history in tqdm(histories, desc="extracting obs and pos")
     ]
 
-    initial_observations, later_observations = (
-        zip(*obs_pairs) if obs_pairs else ([], [])
+    initial_observations = np.array([pair[0] for pair in obs_pairs])
+    initial_observations = initial_observations.reshape(
+        initial_observations.shape[0] * initial_observations.shape[1],
+        initial_observations.shape[2],
+    )
+
+    later_observations = np.array([pair[1] for pair in obs_pairs])
+    later_observations = later_observations.reshape(
+        later_observations.shape[0] * later_observations.shape[1],
+        later_observations.shape[2],
     )
 
     return np.array(initial_observations), np.array(later_observations)
@@ -80,14 +89,20 @@ def extract_pairs_from_histories(histories, agent_name, m, n=None):
 
 def get_torch_from_algo(algo, agent, memory):
     env = env_creator()
-    torch_path = f".{env.metadata['name']}/{memory}/{agent}/torch.pt"
-    if os.path.exists(torch_path):
-        policy_net = torch.load(torch_path)
+    path_ider = f".{env.metadata['name']}/{memory}/{agent}/torch/*.pt"
+    torch_path = glob.glob(path_ider)
+    if not torch_path == []:
+        policy_net = torch.load(torch_path[0], weights_only=False)
     else:
-        algo.get_policy(policy_id=agent).export_model(export_dir=torch_path)
-        policy_net = torch.load(torch_path)
+        algo.get_policy(policy_id=agent).export_model(
+            export_dir="/".join(path_ider.split("/")[:-1])
+        )
+        torch_path = glob.glob(path_ider)
+        policy_net = torch.load(torch_path[0], weights_only=False)
 
-    net_with_softmax = nn.Sequential(*policy_net, nn.Softmax())
+    net_with_softmax = nn.Sequential(
+        policy_net._hidden_layers, policy_net._logits, nn.Softmax(dim=0)
+    )
 
     return net_with_softmax
 
@@ -101,14 +116,11 @@ def compute(
     explainer_extras="none",
     memory="no_memory",
     seed=42,
-    device=None,
+    device=torch.device("cpu"),
 ):
     assert extras in ["none", "one-hot", "action"]
     assert explainer_extras in ["none", "ig", "shap"]
     assert memory in ["no_memory", "lstm", "attention"]
-
-    if not device:
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
     if extras == "one-hot":
         feature_names.extend(act_dict.values())
@@ -126,26 +138,45 @@ def compute(
 
     algo = PPO.from_checkpoint(policy_path)
 
-    paths = glob.glob(f".{env_name}/{memory}/prediction_data_part[0-9].xz")
-    if len(paths) < 10:
-        print(paths)
-        paths = get_future_data(policy_path, memory, seed=921, finished=len(paths))
+    net = get_torch_from_algo(algo, agent, memory)
 
-    X = []
-    y = []
-    for path in paths:
-        with lzma.open(path, "rb") as f:
-            seq = f.read()
+    ray.shutdown()
 
-        partX, party = extract_pairs_from_histories(seq, agent, 10)
-        X.extend(partX)
-        y.extend(party)
+    if not os.path.exists(
+        f".{env_name}/{memory}/{agent}/pred_data/prediction_data.pkl"
+    ):
+        print("Prediction data not found, creating...")
+        paths = glob.glob(f".{env_name}/{memory}/prediction_data_part[0-9].xz")
+        # if len(paths) < 10:
+        #     paths = get_future_data(net, memory, seed=921, finished=paths)
+
+        X = []
+        y = []
+        for path in paths:
+            with lzma.open(path, "rb") as f:
+                seq = pickle.load(f)
+
+            partX, party = extract_pairs_from_histories(seq, agent, 10)
+            X.extend(partX)
+            y.extend(party)
+
+        X = torch.Tensor(np.array(X))
+        y = torch.Tensor(np.array(y))
+        with open(
+            f".{env_name}/{memory}/{agent}/pred_data/prediction_data.pkl", "wb"
+        ) as f:
+            pickle.dump((X, y), f)
+    else:
+        with open(
+            f".{env_name}/{memory}/{agent}/pred_data/prediction_data.pkl", "rb"
+        ) as f:
+            X, y = pickle.load(f)
 
     if extras != "none":
         if not os.path.exists(
             f".{env_name}/{memory}/{agent}/pred_data/prediction_data_action.pkl"
         ):
-            add_action(X, algo, agent, memory)
+            add_action(X, net, agent, memory)
             with open(
                 f".{env_name}/{memory}/{agent}/pred_data/prediction_data_action.pkl",
                 "rb",
@@ -164,11 +195,10 @@ def compute(
         if not os.path.exists(
             f".{env.metadata['name']}/{memory}/{agent}/pred_data/prediction_data_ig_{extras}.pkl"
         ):
-            policy_net = get_torch_from_algo(algo, agent, memory)
-            ig = IntegratedGradients(policy_net)
+            ig = IntegratedGradients(net)
             if not os.path.exists(f".{env_name}/{memory}/{agent}/.baseline_future.pt"):
                 baseline = create_baseline(
-                    algo, agent, device, steps_per_cycle=100, seed=seed
+                    net, agent, device, steps_per_cycle=100, seed=seed
                 )
                 torch.save(
                     baseline, f".{env_name}/{memory}/{agent}/.baseline_future.pt"
@@ -187,7 +217,7 @@ def compute(
                 return_convergence_delta=False,
             )
             X = add_ig(
-                algo,
+                net,
                 agent,
                 memory,
                 X,
@@ -223,17 +253,15 @@ def compute(
                 ) as f:
                     pickle.dump(X, f)
             else:
-                num_acts = env.action_space.n
+                num_acts = env.action_space(agent + "_0").n
 
                 expl = [
-                    kernel_explainer(
-                        algo, agent, i, memory, device, seed=372894 * (i + 1)
-                    )
+                    kernel_explainer(net, agent, i, device, seed=372894 * (i + 1))
                     for i in range(num_acts)
                 ]
 
                 X = add_shap(
-                    algo,
+                    net,
                     agent,
                     memory,
                     X,
@@ -249,7 +277,7 @@ def compute(
             ) as f:
                 X = pickle.load(f)
 
-    net = nn.Sequential(
+    pred_net = nn.Sequential(
         nn.Linear(len(X[0]), 64),
         nn.Tanh(),
         nn.Linear(64, 64),
@@ -260,29 +288,29 @@ def compute(
     if not os.path.exists(
         f".{env_name}/{memory}/{agent}/pred_models/pred_model_{extras}_{explainer_extras}.pt"
     ):
-        net = future_sight(
+        pred_net = future_sight(
             agent,
             memory,
             device,
-            net,
+            pred_net,
             X,
             y,
             extras=extras,
             explainer_extras=explainer_extras,
         )
-        net.eval()
+        pred_net.eval()
     else:
         print(
             f".{env_name}/{memory}/{agent}/pred_models/pred_model_{extras}_{explainer_extras}.pt",
         )
-        net.load_state_dict(
+        pred_net.load_state_dict(
             torch.load(
                 f".{env_name}/{memory}/{agent}/pred_models/pred_model_{extras}_{explainer_extras}.pt",
                 weights_only=True,
                 map_location=device,
             )
         )
-        net.eval()
+        pred_net.eval()
 
     with torch.no_grad():
         criterion = nn.MSELoss()
@@ -297,10 +325,9 @@ def compute(
                 X_test, y_test = pickle.load(f)
         else:
             print("Test data not found, creating...")
-            if not os.path.exists(
-                f".{env_name}/{memory}/{agent}/pred_data/prediction_test_data.pkl"
-            ):
-                path = get_future_data(
+            path = f".{env_name}/{memory}/prediction_test_data.xz"
+            if not os.path.exists(f".{env_name}/{memory}/prediction_test_data.xz"):
+                get_future_data(
                     policy_path,
                     memory,
                     amount_cycles=10000,
@@ -309,16 +336,24 @@ def compute(
                     seed=483927,
                 )
 
+            if not os.path.exists(
+                f".{env_name}/{memory}/{agent}/pred_data/prediction_test_data.pkl"
+            ):
                 with lzma.open(path, "rb") as f:
-                    seq = f.read()
+                    seq = pickle.load(f)
 
                 X_test, y_test = extract_pairs_from_histories(seq, agent, 10)
 
+                os.makedirs(
+                    f".{env_name}/{memory}/{agent}/pred_data",
+                    exist_ok=True,
+                )
                 with open(
                     f".{env_name}/{memory}/{agent}/pred_data/prediction_test_data.pkl",
                     "wb",
                 ) as f:
                     pickle.dump((X_test, y_test), f)
+
             else:
                 with open(
                     f".{env_name}/{memory}/{agent}/pred_data/prediction_test_data.pkl",
@@ -327,18 +362,17 @@ def compute(
                     X_test, y_test = pickle.load(f)
 
             if not extras == "none":
-                X_test = add_action(X_test, algo, agent, memory, save=False)
+                X_test = add_action(X_test, net, agent, memory, save=False)
                 if extras == "one-hot":
                     X_test = one_hot_action(X_test)
 
             if explainer_extras == "ig":
-                policy_net = get_torch_from_algo(algo, agent, memory)
-                ig = IntegratedGradients(policy_net)
+                ig = IntegratedGradients(net)
                 if not os.path.exists(
                     f".{env_name}/{memory}/{agent}/.baseline_future.pt"
                 ):
                     baseline = create_baseline(
-                        algo, agent, device, steps_per_cycle=100, seed=seed
+                        net, agent, device, steps_per_cycle=100, seed=seed
                     )
                     torch.save(
                         baseline, f".{env_name}/{memory}/{agent}/.baseline_future.pt"
@@ -349,6 +383,7 @@ def compute(
                         map_location=device,
                         weights_only=True,
                     )
+                    print(f"{baseline.dtype=}")
                 ig_partial = partial(
                     ig.attribute,
                     baselines=baseline,
@@ -357,7 +392,7 @@ def compute(
                 )
 
                 X_test = add_ig(
-                    algo,
+                    net,
                     agent,
                     memory,
                     X_test,
@@ -382,18 +417,15 @@ def compute(
                         for i in tqdm(range(len(X_test)))
                     ]
                 else:
-                    tempenv = ss.black_death_v3(env)
-                    num_acts = tempenv.action_space.n
+                    num_acts = env.action_space(agent + "_0").n
 
                     expl = [
-                        kernel_explainer(
-                            env, policy_path, 0, i, device, seed=372894 * (i + 1)
-                        )
+                        kernel_explainer(net, agent, i, device, seed=372894 * (i + 1))
                         for i in range(num_acts)
                     ]
 
                     X_test = add_shap(
-                        algo,
+                        net,
                         agent,
                         memory,
                         X_test,
@@ -410,7 +442,7 @@ def compute(
 
         X_test = torch.Tensor(numpyfy(X_test)).to(device)
         y_test = torch.Tensor(numpyfy(y_test)).to(device)
-        test_outputs = net(X_test)
+        test_outputs = pred_net(X_test)
 
         test_loss = criterion(test_outputs, y_test).item()
         print(
@@ -421,26 +453,67 @@ def compute(
         avg_distance = np.mean(distances)
         max_distance = np.max(distances)
 
+    expl = [
+        kernel_explainer(pred_net, agent, i, device, seed=372894 * (i + 1))
+        for i in range(len(y[0]))
+    ]
+
+    indices = torch.randperm(len(X_test))[:50]
+    for i in range(len(y[0])):
+        make_plots(
+            expl,
+            X_test[indices],
+            agent,
+            memory,
+            extras,
+            explainer_extras,
+            i,
+        )
+
     return (test_loss, avg_distance, max_distance)
 
 
-def make_plots(explainer, X, filename):
+def make_plots(explainer, X, agent, memory, extras, explainer_extras, target):
     if not isinstance(X, np.ndarray):
         X = numpyfy(X)
 
-    coordinate = 0
-    coordinate_names = ["x", "y"]
+    env = env_creator()
+    feature_names = env.feature_names
 
-    shap_plot(
-        X,
-        explainer,
-        filename,
-        feature_names,
-        coordinate_names[coordinate],
-    )
+    if extras == "action":
+        feature_names.append("action")
+    elif extras == "one-hot":
+        feature_names.extend(env.act_dict.keys())
+
+    if explainer_extras != "none":
+        feature_names.extend(
+            [
+                feature_name + " " + explainer_extras
+                for feature_name in env.feature_names
+            ]
+        )
+    if isinstance(explainer, list):
+        for expl in explainer:
+            shap_plot(
+                agent,
+                memory,
+                X,
+                expl,
+                feature_names,
+                target,
+            )
+    else:
+        shap_plot(
+            agent,
+            memory,
+            X,
+            explainer,
+            feature_names,
+            target,
+        )
 
 
-def run_compare(policy_path, agent, memory, feature_names, act_dict):
+def run_compare(policy_path, agent, memory, feature_names, act_dict, device):
     extras = ["none", "action", "one-hot"]
     explainer_extras = ["none", "ig", "shap"]
 
@@ -456,6 +529,7 @@ def run_compare(policy_path, agent, memory, feature_names, act_dict):
                 extras=extra,
                 explainer_extras=expl,
                 memory=memory,
+                device=device,
             )
             for k, out in enumerate(outs):
                 table[i, j, k] = out
@@ -465,97 +539,3 @@ def run_compare(policy_path, agent, memory, feature_names, act_dict):
         pickle.dump(table, f)
 
     return table
-
-
-# if __name__ == "__main__":
-#     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-#     seed = 42
-#     # Superseeding, might be unnecessary
-#     np.random.seed(seed)
-#     random.seed(seed)
-#
-#     parser = argparse.ArgumentParser(description="Simulation")
-#     parser.add_argument(
-#         "-e",
-#         "--env",
-#         type=str,
-#         help="Which environment to use",
-#         default="spread",
-#     )
-#     parser.add_argument(
-#         "-r",
-#         "--render",
-#         type=str,
-#         help="Render mode, default None",
-#         default=None,
-#     )
-#
-#     args = parser.parse_args()
-#
-#     if args.env == "spread":
-#         env_fn = simple_spread_v3
-#
-#         env_kwargs = dict(
-#             N=3,
-#             local_ratio=0.5,
-#             max_cycles=25,
-#             continuous_actions=False,
-#         )
-#
-#         feature_names = [
-#             "vel x",
-#             "vel y",
-#             "pos x",
-#             "pos y",
-#             "landmark 1 x",
-#             "landmark 1 y",
-#             "landmark 2 x",
-#             "landmark 2 y",
-#             "landmark 3 x",
-#             "landmark 3 y",
-#             "agent 2 x",
-#             "agent 2 y",
-#             "agent 3 x",
-#             "agent 3 y",
-#             "comms 1",
-#             "comms 2",
-#             "comms 3",
-#             "comms 4",
-#         ]
-#         act_dict = {
-#             0: "no action",
-#             1: "move left",
-#             2: "move right",
-#             3: "move down",
-#             4: "move up",
-#         }
-#     elif args.env == "kaz":
-#         env_fn = knights_archers_zombies_v10
-#
-#         env_kwargs = dict(
-#             spawn_rate=6,
-#             num_archers=2,
-#             num_knights=2,
-#             max_zombies=10,
-#             max_arrows=10,
-#             max_cycles=900,
-#             vector_state=True,
-#         )
-#         feature_names = None
-#         act_dict = None
-#     else:
-#         print("Invalid env entered")
-#         exit(0)
-#
-#     env = env_fn.parallel_env(render_mode=args.render, **env_kwargs)
-#     try:
-#         policy_path = max(
-#             glob.glob(f".{str(env.metadata['name'])}/*.zip"),
-#             key=os.path.getctime,
-#         )
-#         print(policy_path)
-#     except ValueError:
-#         print("Policy not found in " + f".{str(env.metadata['name'])}/*.zip")
-#         exit(0)
-#
-#     run_compare(policy_path, feature_names, act_dict)
